Refactor: Move createStreamingAdWrapper to a separate file for better modularity
