+++ text.pollinations.ai/package.json
@@ -26,6 +26,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "eventsource": "^3.0.2",
+    "eventsource-parser": "^3.0.2",
     "express": "^4.20.0",
     "express-rate-limit": "^7.5.0",
     "jsonwebtoken": "^9.0.2",
